+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `id_translation.logging.EMIT_LOGGED_WARNINGS`; set to `False` to disable warnings that are emitted as logs.
+
 ## [1.2.1] - 2026-05-09
 
 ### Fixed